+++ catalog/service/impl/src/test/java/org/projectnessie/catalog/service/impl/TestCatalogServiceImpl.java
@@ -20,20 +20,26 @@ import static java.util.concurrent.TimeUnit.MINUTES;
 import static org.assertj.core.api.InstanceOfAssertFactories.STRING;
 import static org.projectnessie.api.v2.params.ParsedReference.parsedReference;
 import static org.projectnessie.catalog.service.api.SnapshotReqParams.forSnapshotHttpReq;
+import static org.projectnessie.model.CommitMeta.fromMessage;
 import static org.projectnessie.model.Content.Type.ICEBERG_TABLE;
 import static org.projectnessie.services.authz.Check.CheckType.COMMIT_CHANGE_AGAINST_REFERENCE;
 import static org.projectnessie.services.authz.Check.CheckType.READ_ENTITY_VALUE;
 import static org.projectnessie.services.authz.Check.CheckType.UPDATE_ENTITY;
 import static org.projectnessie.services.authz.Check.CheckType.VIEW_REFERENCE;
 
+import java.io.InputStream;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.assertj.core.api.AbstractThrowableAssert;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.projectnessie.api.v2.params.ParsedReference;
 import org.projectnessie.catalog.formats.iceberg.meta.IcebergJson;
 import org.projectnessie.catalog.formats.iceberg.meta.IcebergTableMetadata;
@@ -43,19 +49,113 @@ import org.projectnessie.catalog.service.api.CatalogCommit;
 import org.projectnessie.catalog.service.api.CatalogService;
 import org.projectnessie.catalog.service.api.SnapshotReqParams;
 import org.projectnessie.catalog.service.api.SnapshotResponse;
+import org.projectnessie.error.NessieReferenceConflictException;
+import org.projectnessie.model.Branch;
 import org.projectnessie.model.Content;
 import org.projectnessie.model.ContentKey;
 import org.projectnessie.model.IcebergTable;
+import org.projectnessie.model.IcebergView;
+import org.projectnessie.model.Operation;
 import org.projectnessie.model.Reference;
+import org.projectnessie.objectstoragemock.Bucket;
+import org.projectnessie.objectstoragemock.MockObject;
 import org.projectnessie.services.authz.AbstractBatchAccessChecker;
 import org.projectnessie.services.authz.AccessCheckException;
 import org.projectnessie.services.authz.Check;
 import org.projectnessie.services.authz.Check.CheckType;
 import org.projectnessie.storage.uri.StorageUri;
+import software.amazon.awssdk.services.s3.model.S3Exception;
 
 public class TestCatalogServiceImpl extends AbstractCatalogService {
 
   /** Check that a non-modifying catalog-commit is a no-op. */
+  @Test
+  public void cleanupAfterNessieCommitFailure() throws Exception {
+    Reference main = api.getReference().refName("main").get();
+    ContentKey key1 = ContentKey.of("mytable");
+    ContentKey key2 = ContentKey.of("othertable");
+
+    AtomicReference<List<String>> storedLocations = new AtomicReference<>(new ArrayList<>());
+
+    interceptingBucket.setUpdater(
+        (k, m) -> {
+          List<String> l = storedLocations.get();
+          l.add(k);
+          storedLocations.set(l);
+          return Optional.empty();
+        });
+
+    api.commitMultipleOperations()
+        .branch((Branch) main)
+        .commitMeta(fromMessage("break next commit"))
+        .operation(Operation.Put.of(key1, IcebergView.of("meta", 1, 2)))
+        .commitWithResponse();
+
+    soft.assertThatThrownBy(() -> commitMultiple(main, key1, key2))
+        .isInstanceOf(ExecutionException.class)
+        .cause()
+        .isInstanceOf(RuntimeException.class)
+        .hasCauseInstanceOf(NessieReferenceConflictException.class);
+    soft.assertThat(storedLocations.get()).hasSize(2);
+    soft.assertThat(heapStorageBucket.objects()).isEmpty();
+  }
+
+  @ParameterizedTest
+  @ValueSource(ints = {0, 1, 2})
+  public void cleanupAfterObjectIoFailure(int after) throws Exception {
+    Reference main = api.getReference().refName("main").get();
+    ContentKey key1 = ContentKey.of("mytable1");
+    ContentKey key2 = ContentKey.of("mytable2");
+    ContentKey key3 = ContentKey.of("mytable3");
+    ContentKey key4 = ContentKey.of("mytable4");
+
+    AtomicReference<List<String>> storedLocations = new AtomicReference<>(new ArrayList<>());
+    AtomicReference<List<String>> failedLocations = new AtomicReference<>(new ArrayList<>());
+
+    interceptingBucket.setUpdater(
+        (k, m) -> {
+          List<String> l = storedLocations.get();
+          if (l.size() == after) {
+            l = failedLocations.get();
+            l.add(k);
+            failedLocations.set(l);
+            return Optional.of(
+                new Bucket.ObjectUpdater() {
+                  @Override
+                  public Bucket.ObjectUpdater append(long position, InputStream data) {
+                    return this;
+                  }
+
+                  @Override
+                  public Bucket.ObjectUpdater flush() {
+                    return this;
+                  }
+
+                  @Override
+                  public Bucket.ObjectUpdater setContentType(String contentType) {
+                    return this;
+                  }
+
+                  @Override
+                  public MockObject commit() {
+                    throw new UnsupportedOperationException("Injected Object Storage Failure");
+                  }
+                });
+          }
+          l.add(k);
+          storedLocations.set(l);
+          return Optional.empty();
+        });
+
+    soft.assertThatThrownBy(() -> commitMultiple(main, key1, key2, key3, key4))
+        .isInstanceOf(ExecutionException.class)
+        .cause()
+        .isInstanceOf(S3Exception.class);
+    soft.assertThat(storedLocations.get()).hasSize(after);
+    soft.assertThat(failedLocations.get()).hasSize(4 - after);
+    soft.assertThat(heapStorageBucket.objects()).isEmpty();
+  }
+
   @Test
   public void noCommitOps() throws Exception {
     Reference main = api.getReference().refName("main").get();
@@ -71,6 +171,21 @@ public class TestCatalogServiceImpl extends AbstractCatalogService {
   }
 
   /** Verifies that a single table create catalog-commit passes. */
+  @Test
+  public void twoTableCreates() throws Exception {
+    Reference main = api.getReference().refName("main").get();
+    ContentKey key1 = ContentKey.of("mytable1");
+    ContentKey key2 = ContentKey.of("mytable2");
+
+    ParsedReference committed = commitMultiple(main, key1, key2);
+
+    Reference afterCommit = api.getReference().refName("main").get();
+    soft.assertThat(afterCommit)
+        .isNotEqualTo(main)
+        .extracting(Reference::getName, Reference::getHash)
+        .containsExactly(committed.name(), committed.hashWithRelativeSpec());
+  }
+
   @Test
   public void singleTableCreate() throws Exception {
     Reference main = api.getReference().refName("main").get();