+++ CHANGELOG.md
@@ -17,6 +17,7 @@ as necessary. Empty sections will not end in the release notes.
 - Catalog/GCS: Support using the default application credentials
 - Catalog/S3: Allow custom key+trust stores
 - Catalog: Check privileges earlier
+- Catalog: cleanup non-committed metadata objects
 
 ### Changes